refactor: move L2BlockId to primitives
