+++ crates/primitives/src/l2.rs
@@ -1,0 +1,25 @@
+use arbitrary::Arbitrary;
+use borsh::{BorshDeserialize, BorshSerialize};
+use serde::{Deserialize, Serialize};
+
+use crate::{buf::Buf32, impl_buf_wrapper};
+
+/// ID of an L2 block, usually the hash of its root header.
+#[derive(
+    Copy,
+    Clone,
+    Eq,
+    Default,
+    PartialEq,
+    Ord,
+    PartialOrd,
+    Hash,
+    Arbitrary,
+    BorshSerialize,
+    BorshDeserialize,
+    Serialize,
+    Deserialize,
+)]
+pub struct L2BlockId(Buf32);
+
+impl_buf_wrapper!(L2BlockId, Buf32, 32);
